🔨 [refactor] Article, Reply Domain 디렉토리로 리팩토링
