--- src/main/java/com/example/umc8th/article/dto/ArticleRequestDTO.java
@@ -1,12 +1,0 @@
-package com.example.umc8th.article.dto;
-
-import lombok.Getter;
-
-public class ArticleRequestDTO {
-
-    @Getter
-    public static class CreateArticleDTO {
-        private String title;
-        private String content;
-    }
-}
