--- src/main/java/com/example/umc8th/article/dto/ArticleResponseDTO.java
@@ -1,29 +1,0 @@
-package com.example.umc8th.article.dto;
-
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-
-public class ArticleResponseDTO {
-
-    @Builder
-    @Data
-    public static class ArticleDTO {
-        private Long articleId;
-        private String content;
-        private String title;
-        private LocalDateTime createdAt;
-        private LocalDateTime updatedAt;
-    }
-
-    @Getter
-    @Builder
-    public static class ArticleListDTO {
-        private List<ArticleDTO> articles;
-    }
-}
