--- src/main/java/com/example/umc8th/article/repository/ArticleRepository.java
@@ -1,8 +1,0 @@
-package com.example.umc8th.article.repository;
-
-import com.example.umc8th.article.entity.Article;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ArticleRepository extends JpaRepository<Article, Long> {
-
-}
