--- src/main/java/com/example/umc8th/article/service/command/ArticleCommandService.java
@@ -1,8 +1,0 @@
-package com.example.umc8th.article.service.command;
-
-import com.example.umc8th.article.dto.ArticleRequestDTO;
-import com.example.umc8th.article.dto.ArticleResponseDTO;
-
-public interface ArticleCommandService {
-    ArticleResponseDTO.ArticleDTO createArticle(ArticleRequestDTO.CreateArticleDTO dto);
-}
