--- src/main/java/com/example/umc8th/article/service/command/ArticleCommandServiceImpl.java
@@ -1,23 +1,0 @@
-package com.example.umc8th.article.service.command;
-
-import com.example.umc8th.article.converter.ArticleConverter;
-import com.example.umc8th.article.dto.ArticleRequestDTO;
-import com.example.umc8th.article.dto.ArticleResponseDTO;
-import com.example.umc8th.article.entity.Article;
-import com.example.umc8th.article.repository.ArticleRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class ArticleCommandServiceImpl implements ArticleCommandService {
-    private final ArticleRepository articleRepository;
-
-    @Override
-    public ArticleResponseDTO.ArticleDTO createArticle(ArticleRequestDTO.CreateArticleDTO dto) {
-        Article article = ArticleConverter.toArticle(dto);
-        return ArticleConverter.toArticleDTO(articleRepository.save(article));
-    }
-}
