--- src/main/java/com/example/umc8th/article/service/query/ArticleQueryService.java
@@ -1,8 +1,0 @@
-package com.example.umc8th.article.service.query;
-
-import com.example.umc8th.article.dto.ArticleResponseDTO;
-
-public interface ArticleQueryService {
-   ArticleResponseDTO.ArticleDTO getArticle(Long articleId);
-   ArticleResponseDTO.ArticleListDTO getArticles();
-}
