+++ src/main/java/com/example/umc8th/domain/reply/dto/ReplyRequestDTO.java
@@ -1,0 +1,12 @@
+package com.example.umc8th.domain.reply.dto;
+
+import lombok.Getter;
+
+
+public class ReplyRequestDTO {
+
+    @Getter
+    public static class CreateReplyDTO{
+        private String content;
+    }
+}
