+++ src/main/java/com/example/umc8th/domain/reply/service/query/ReplyQueryService.java
@@ -1,0 +1,7 @@
+package com.example.umc8th.domain.reply.service.query;
+
+import com.example.umc8th.domain.reply.dto.ReplyResponseDTO;
+
+public interface ReplyQueryService {
+    ReplyResponseDTO.ReplyListDTO getReplyList(Long articleId);
+}
