+++ src/main/java/com/example/umc8th/domain/reply/service/query/ReplyQueryServiceImpl.java
@@ -1,0 +1,31 @@
+package com.example.umc8th.domain.reply.service.query;
+
+import com.example.umc8th.domain.article.entity.Article;
+import com.example.umc8th.domain.article.exception.ArticleException;
+import com.example.umc8th.domain.article.exception.code.ArticleErrorCode;
+import com.example.umc8th.domain.article.repository.ArticleRepository;
+import com.example.umc8th.domain.reply.converter.ReplyConverter;
+import com.example.umc8th.domain.reply.dto.ReplyResponseDTO;
+import com.example.umc8th.domain.reply.entity.Reply;
+import com.example.umc8th.domain.reply.repository.ReplyRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class ReplyQueryServiceImpl implements ReplyQueryService {
+    private final ArticleRepository articleRepository;
+    private final ReplyRepository replyRepository;
+
+    @Override
+    public ReplyResponseDTO.ReplyListDTO getReplyList(Long articleId) {
+        Article article = articleRepository.findById(articleId).orElseThrow(() ->
+                new ArticleException(ArticleErrorCode.NOT_FOUND_404));
+        List<Reply> replies = replyRepository.findAllByArticle(article);
+        return ReplyConverter.toReplyListDTO(replies);
+    }
+}
