--- src/main/java/com/example/umc8th/reply/converter/ReplyConverter.java
@@ -1,38 +1,0 @@
-package com.example.umc8th.reply.converter;
-
-import com.example.umc8th.article.entity.Article;
-import com.example.umc8th.reply.dto.ReplyRequestDTO;
-import com.example.umc8th.reply.dto.ReplyResponseDTO;
-import com.example.umc8th.reply.entity.Reply;
-
-import java.util.List;
-
-public class ReplyConverter {
-
-    // CreateReplyDTO -> Reply
-    public static Reply toReply(ReplyRequestDTO.CreateReplyDTO dto, Article article) {
-        return Reply.builder()
-                .content(dto.getContent())
-                .article(article)
-                .build();
-    }
-
-    // Reply -> ReplyDTO
-    public static ReplyResponseDTO.ReplyDTO toReplyDTO(Reply reply) {
-        return ReplyResponseDTO.ReplyDTO.builder()
-                .content(reply.getContent())
-                .articleId(reply.getArticle().getId())
-                .createdAt(reply.getCreatedAt())
-                .updatedAt(reply.getUpdatedAt())
-                .build();
-    }
-
-    // List<Reply> -> ReplyListDTO
-    public static ReplyResponseDTO.ReplyListDTO toReplyListDTO(List<Reply> reply) {
-        return ReplyResponseDTO.ReplyListDTO.builder()
-                .replies(reply.stream()
-                        .map(ReplyConverter::toReplyDTO)
-                        .toList())
-                .build();
-    }
-}
