--- src/main/java/com/example/umc8th/reply/dto/ReplyRequestDTO.java
@@ -1,12 +1,0 @@
-package com.example.umc8th.reply.dto;
-
-import lombok.Getter;
-
-
-public class ReplyRequestDTO {
-
-    @Getter
-    public static class CreateReplyDTO{
-        private String content;
-    }
-}
