--- src/main/java/com/example/umc8th/reply/service/command/ReplyCommandService.java
@@ -1,8 +1,0 @@
-package com.example.umc8th.reply.service.command;
-
-import com.example.umc8th.reply.dto.ReplyRequestDTO;
-import com.example.umc8th.reply.dto.ReplyResponseDTO;
-
-public interface ReplyCommandService {
-    ReplyResponseDTO.ReplyDTO createReply(ReplyRequestDTO.CreateReplyDTO dto, Long articleId);
-}
