--- src/main/java/com/example/umc8th/reply/service/query/ReplyQueryService.java
@@ -1,7 +1,0 @@
-package com.example.umc8th.reply.service.query;
-
-import com.example.umc8th.reply.dto.ReplyResponseDTO;
-
-public interface ReplyQueryService {
-    ReplyResponseDTO.ReplyListDTO getReplyList(Long articleId);
-}
